implementation of the date_time column; deletion of console.log from the code; refactoring of DateUtilService;
the date_time column will replace the date and time columns that were separate. the getZonedDate method of DateUtilService always aims to convert the times that come from the bank in UTC to the client's timezone. The unification of time and date will facilitate this process in all services that will display the time of the consultation or batches of consultants' availability